Refactors list-component and projectable-list-component
Introduces a base component for list components and a presentation component for list components.
- Also injects a scrolling-table-component by its dedicated service (#1217)
- Also saves a search query (#1221)
