+++ client/src/app/site/pages/meetings/modules/meetings-component-collector/projectable-list/components/projectable-list/projectable-list.component.ts
@@ -1,11 +1,13 @@
-import { ChangeDetectionStrategy, ChangeDetectorRef, Component, Input, ViewEncapsulation } from '@angular/core';
+import { ChangeDetectionStrategy, ChangeDetectorRef, Component, Input } from '@angular/core';
 import { Permission } from 'src/app/domain/definitions/permission';
 import { BaseViewModel } from 'src/app/site/base/base-view-model';
 import { HasListOfSpeakers, hasListOfSpeakers } from 'src/app/site/pages/meetings/pages/agenda';
 import { BaseProjectableViewModel } from 'src/app/site/pages/meetings/view-models/base-projectable-model';
 import { OperatorService } from 'src/app/site/services/operator.service';
 import { ViewPortService } from 'src/app/site/services/view-port.service';
-import { ColumnRestriction, ListComponent } from 'src/app/ui/modules/list/components';
+import { ColumnRestriction, ListComponent } from 'src/app/ui/modules/list';
+import { BaseListComponent } from 'src/app/ui/modules/list/base/base-list.component';
+import { ScrollingTableManageService } from 'src/app/ui/modules/scrolling-table';
 import {
     END_POSITION,
     START_POSITION
@@ -25,10 +27,9 @@ export interface CssClassDefinition {
     templateUrl: `./projectable-list.component.html`,
     styleUrls: [`./projectable-list.component.scss`],
     changeDetection: ChangeDetectionStrategy.OnPush,
-    encapsulation: ViewEncapsulation.None,
     providers: [{ provide: ListComponent, useExisting: ProjectableListComponent }]
 })
-export class ProjectableListComponent<V extends BaseViewModel | BaseProjectableViewModel> extends ListComponent<V> {
+export class ProjectableListComponent<V extends BaseViewModel | BaseProjectableViewModel> extends BaseListComponent<V> {
     public readonly END_POSITION = END_POSITION;
     public readonly START_POSITION = START_POSITION;
 
@@ -78,10 +79,11 @@ export class ProjectableListComponent<V extends BaseViewModel | BaseProjectableV
     public constructor(
         vp: ViewPortService,
         cd: ChangeDetectorRef,
+        scrollingTableManager: ScrollingTableManageService,
         private operator: OperatorService,
         private service: ProjectableListService
     ) {
-        super(vp, cd);
+        super(vp, cd, scrollingTableManager);
     }
 
     public _getSpeakerButtonObject(viewModel: V): (BaseViewModel & HasListOfSpeakers) | null {
