+++ client/src/app/site/pages/meetings/modules/projector/modules/slides/components/motions/modules/motion-slide/motion-slide.module.ts
@@ -1,13 +1,12 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { OpenSlidesTranslationModule } from 'src/app/site/modules/translations';
-import { MotionsCommonServiceModule } from 'src/app/site/pages/meetings/pages/motions/services/common/motions-service.module';
 import { PipesModule } from 'src/app/ui/pipes';
 
 import { SlideToken } from '../../../../definitions';
 import { MotionSlideComponent } from './components/motion-slide/motion-slide.component';
 @NgModule({
-    imports: [CommonModule, PipesModule, OpenSlidesTranslationModule.forChild(), MotionsCommonServiceModule],
+    imports: [CommonModule, PipesModule, OpenSlidesTranslationModule.forChild()],
     declarations: [MotionSlideComponent],
     providers: [{ provide: SlideToken.token, useValue: MotionSlideComponent }]
 })
