+++ client/src/app/site/pages/meetings/pages/motions/pages/motion-call-list/motion-call-list.module.ts
@@ -10,7 +10,6 @@ import { HeadBarModule } from 'src/app/ui/modules/head-bar';
 import { SortingModule } from 'src/app/ui/modules/sorting';
 
 import { MotionCategoryCommonServiceModule, TagCommonServiceModule } from '../../modules';
-import { MotionsCommonServiceModule } from '../../services/common/motions-service.module';
 import { MotionsExportModule } from '../../services/export/motions-export.module';
 import { MotionCallListComponent } from './components/motion-call-list/motion-call-list.component';
 import { MotionCallListRoutingModule } from './motion-call-list-routing.module';
@@ -20,7 +19,6 @@ import { MotionCallListRoutingModule } from './motion-call-list-routing.module';
     imports: [
         CommonModule,
         MotionCallListRoutingModule,
-        MotionsCommonServiceModule,
         MotionsExportModule,
         MotionCategoryCommonServiceModule,
         TagCommonServiceModule,
