+++ client/src/app/site/pages/meetings/pages/motions/services/common/motions-service.module.ts
@@ -1,8 +1,6 @@
 import { NgModule } from '@angular/core';
 
-import { ChangeRecommendationsModule } from '../../modules/change-recommendations';
-
 @NgModule({
-    imports: [ChangeRecommendationsModule]
+    imports: []
 })
 export class MotionsCommonServiceModule {}
