+++ client/src/app/ui/modules/list/components/list/list.component.scss
@@ -1,26 +1,26 @@
-.list-view-table-wrapper {
-    display: flex;
-    flex-flow: column;
-    height: 100%;
+// .list-view-table-wrapper {
+//     display: flex;
+//     flex-flow: column;
+//     height: 100%;
 
-    .sort-filter-bar {
-        flex: 0 1 auto;
-        display: block;
-    }
+//     .sort-filter-bar {
+//         flex: 0 1 auto;
+//         display: block;
+//     }
 
-    .vscroll-list-view {
-        flex: 1 1 auto;
-        height: 100%;
-    }
+//     .vscroll-list-view {
+//         flex: 1 1 auto;
+//         height: 100%;
+//     }
 
-    .no-data-placeholder {
-        position: absolute;
-        margin-left: auto;
-        margin-right: auto;
-        top: 5%;
-        left: 0;
-        right: 0;
-        width: 100px; /* Need a specific value to work */
-        font-style: italic;
-    }
-}
+//     .no-data-placeholder {
+//         position: absolute;
+//         margin-left: auto;
+//         margin-right: auto;
+//         top: 5%;
+//         left: 0;
+//         right: 0;
+//         width: 100px; /* Need a specific value to work */
+//         font-style: italic;
+//     }
+// }
