+++ client/src/app/ui/modules/list/components/view-list/view-list.component.spec.ts
@@ -1,0 +1,24 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { ViewListComponent } from './view-list.component';
+
+describe(`ViewListComponent`, () => {
+    let component: ViewListComponent;
+    let fixture: ComponentFixture<ViewListComponent>;
+
+    beforeEach(async () => {
+        await TestBed.configureTestingModule({
+            declarations: [ViewListComponent]
+        }).compileComponents();
+    });
+
+    beforeEach(() => {
+        fixture = TestBed.createComponent(ViewListComponent);
+        component = fixture.componentInstance;
+        fixture.detectChanges();
+    });
+
+    it(`should create`, () => {
+        expect(component).toBeTruthy();
+    });
+});
