+++ client/src/app/ui/modules/list/definitions/column-restriction.ts
@@ -1,0 +1,7 @@
+/**
+ * To hide columns via restriction
+ */
+export interface ColumnRestriction<P = any> {
+    columnName: string;
+    permission: P;
+}
