+++ client/src/app/ui/modules/list/definitions/index.ts
@@ -1,2 +1,4 @@
+export * from './column-restriction';
 export * from './filter-service';
+export * from './search-service';
 export * from './sort-service';
