+++ client/src/app/ui/modules/list/index.ts
@@ -1,1 +1,4 @@
+export * from './base/base-list.component';
+export * from './components/list/list.component';
+export * from './definitions';
 export * from './list.module';
