+++ client/src/app/ui/modules/scrolling-table/components/scrolling-table/scrolling-table.component.ts
@@ -129,11 +129,13 @@ export class ScrollingTableComponent<T extends Partial<Mutable<Identifiable>>>
     }
 
     public ngOnInit(): void {
+        this.manageService.currentScrollingTableComponent = this;
         this.subscriptions.push(this.manageService.cellDefinitionsObservable.subscribe(() => this.cd.markForCheck()));
     }
 
     public override ngOnDestroy(): void {
         super.ngOnDestroy();
+        this.manageService.currentScrollingTableComponent = null;
         this.manageService.clearRegistry();
     }
