+++ client/src/app/ui/modules/scrolling-table/services/scrolling-table-manage.service.ts
@@ -2,6 +2,7 @@ import { TemplatePortal } from '@angular/cdk/portal';
 import { Injectable } from '@angular/core';
 import { BehaviorSubject, Observable } from 'rxjs';
 
+import { ScrollingTableComponent } from '../components/scrolling-table/scrolling-table.component';
 import { ScrollingTableCellDefinition } from '../directives/scrolling-table-cell-definition';
 import { ScrollingTableCellPosition } from '../directives/scrolling-table-cell-position';
 import { ScrollingTableServiceModule } from './scrolling-table-service.module';
@@ -47,6 +48,8 @@ class ScrollingTableCellHandler implements Iterable<ScrollingTableCellDefinition
     providedIn: ScrollingTableServiceModule
 })
 export class ScrollingTableManageService {
+    public currentScrollingTableComponent: ScrollingTableComponent<any> | null = null;
+
     public get cellDefinitionsObservable(): Observable<ScrollingTableCellDefinition[]> {
         return this._cellDefinitionsSubject.asObservable();
     }
